只做了一点的实例化渲染

+++ tests/ecs_test/glsl/model_fs.glsl
@@ -1,5 +1,5 @@
 #version 320 es
-precision lowp float;
+precision highp float;
 
 out vec4 FragColor;
